ADD new Heal Ability script
